+++ changelog.md
@@ -2,3 +2,4 @@
 
 ## 0.0.1 (2022-11-03)
 + setting up package
++ implemented taxonomic distribution sunburst plot using plotly